requirements.txt
web3==6.15.1
eth-account==0.11.0
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
agent0-sdk
